+++ block2/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 * **BREAKING**: Updated `objc2-encode` to `v2.0.0-pre.2`.
+* Updated `ffi` module to `block-sys v0.1.0-beta.1`.
 
 ### Fixed
 * **BREAKING**: Cleaned up `BlockArguments` trait, it is now sealed and a